slowly figuring out how to realise operations
